refactor: 닉네임 유효성 체크 로직 분리
